% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_nlme.r
\name{doxpose}
\alias{doxpose}
\title{Create a grid with 4 common diagnostics plots}
\usage{
doxpose(xp = NULL, modelName = "")
}
\arguments{
\item{xp}{xpose database or NULL to create a new database}

\item{modelName}{Model name}
}
\description{
Create a grid with 4 common diagnostics plots
}
