% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_nlme.r
\name{getObsData}
\alias{getObsData}
\title{getObsData : Updates column names for ID, TIME and DV so input can be passed on to vpc package}
\usage{
getObsData(input, mapFile = "cols1.txt")
}
\arguments{
\item{input}{observation data frame}

\item{mapFile}{column mapping file name}
}
\description{
getObsData : Updates column names for ID, TIME and DV so input can be passed on to vpc package
}
\examples{

input = getObsData(input)



}
