% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_nlme.r
\name{getSimData}
\alias{getSimData}
\title{getSimData : Opens and returns simulation data from an NLME VPC run}
\usage{
getSimData(input, stratifyColumns = "", mapFile = "cols1.txt",
  simFile = "out.txt", modelDir = "./")
}
\arguments{
\item{input}{observation data frame}

\item{stratifyColumns}{comma separted list of covariates to stratify on}

\item{mapFile}{column mapping file name}

\item{simFile}{Filename containing simulation data}

\item{modelDir}{directory where model was executed}
}
\description{
getSimData : Opens and returns simulation data from an NLME VPC run
}
\examples{

simdata = getSimData(input,stratifyColumns="sex",modelDir)



}
