% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_nlme.r
\name{lookupMappedColumn}
\alias{lookupMappedColumn}
\title{lookupMappedColumn looks at column mapping file and maps a column}
\usage{
lookupMappedColumn(columnNames, mapping, colName)
}
\description{
lookupMappedColumn looks at column mapping file and maps a column
}
