% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpose_nlme.r
\name{mapCovariate}
\alias{mapCovariate}
\title{mapCovariate : maps a covariate name to input file column name}
\usage{
mapCovariate(input, covarNames, mapFile = "cols1.txt")
}
\arguments{
\item{input}{observation data frame}

\item{covarNames}{comma separated covariate names}

\item{mapFile}{column mapping file name}
}
\description{
mapCovariate : maps a covariate name to input file column name
}
\examples{

names = mapCovariate(input,"sex,age")



}
